+++ ChangeLog.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 
 ## Version 1.24
 
+- FIX : Missing substitution *04/01/2022* - 1.24.1
 - NEW : Display company on card *02/01/2022* - 1.24.0
 - NEW : Delete cascade from kanban *20/12/2022* - 1.23.0